changed float to double

+++ src/main/java/com/qa/day6/TempConverter.java
@@ -6,20 +6,20 @@ public class TempConverter {
         return (5.0 / 9.0) * (fahrenheit - 32);
     }
 
-    public float convertCelsiusToFahrenheit(int celsius) {
-        return (9 / 5) * (celsius) + 32;
+    public double convertCelsiusToFahrenheit(double celsius) {
+        return (9.0/ 5.0) * (celsius) + 32;
     }
 
-    public float convertKelvinToCelsius(int kelvin) {
+    public double convertKelvinToCelsius(double kelvin) {
         return (kelvin - 273);
     }
 
-    public float convertCelsiusToKelvin(int celsius) {
+    public float convertCelsiusToKelvin(float celsius) {
         return (celsius + 273);
     }
 
-    public float convertKelvinToFahrenheit(int kelvin) {
-        return (9 / 5) * (kelvin - 273) + 32;
+    public double convertKelvinToFahrenheit(double kelvin) {
+        return (9.0 / 5.0) * (kelvin - 273) + 32;
     }
 
     public float convertFahrenheitToKelvin(int fahrenheit) {
